Trying docker-compose for CI
It's difficult and against common docker rules of them to run something
like a database, then run your tests in the same container.

The recommended route is to use multiple containers, which means
compose.  Compose however makes it difficult to get proper exit codes
from one or more test suites, and it doesn't have as nice integration
with Jenkins.  Jenkins always assumes a single container for all of it's
steps.