Started transition back to bootstrap
This project started with bootstrap, then I moved to bulma because I
wanted a more "pure" css driven UI.  The Jquery requirement really put
me off of bootstrap.  After finding lots of bugs with IE 11 in bulma, I
decided to switch to spectre.

Spectre worked well, but compared to previous projects I built with
bootstrap I'm seeing how much I underestimated what bootstrap brought to
the table.

With bootstrap I could build a preliminary UI and only have to tweak a
few small things to get it working well on smaller screens.  I assumed
that it would be the same with spectre, as the documentation has lot of
responsive utilities.  However the lack of a mobile first grid alone
made this very difficult.  I usually rely heavily on the grid to define
the basic outline of a UI, so I was finding that to make my pages mobile
friendly with spectre I was having to rebuild large sections of the
page, or start over completely.

I also found myself re-adding several of the spacing and sizing utilities
that came with bootstrap.  In the end, I'll probably end up adding
several things from Spectre, but my base will be built on bootstrap.

It's looking like bootstrap will be dropping the jquery requirement in
v5.  Until that is completed, I'll be using bootstrap.native.